backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
openpyxl==3.1.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
openai==0.28.1
python-dotenv==1.0.0
langchain==0.0.300
plotly==5.17.0
numpy==1.24.3
matplotlib==3.8.0
python-dotenv==1.0.0
tiktoken==0.5.1